Fold tests into one
